# target/classes/swagger.yaml
openapi: 3.0.0
info:
  title: Back-end Challenge 2021 🏅 - Space Flight News
  description: API Restful com as melhores práticas de desenvolvimento, baseada na API Space Flight News.
  version: 1.0.0
servers:
  - url: http://api.spaceflight.news/v1
    description: Back-end Challenge 2021 🏅 - Space Flight News

paths:

  /pokemon/{id}:
    get:
      tags:
        - Pokemons
      summary: Obter um pokemon pelo seu id
      description: Obter um pokemon pelo seu id
      operationId: findPokemonById
      parameters:
        - name: id
          in: path
          description: ID do pokemon
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':    # status code
          description: sucesso consulta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonDto'

  /pokemon:
    get:
      tags:
        - Pokemons
      summary: Obter uma lista de pokemons
      description: Obter uma lista de pokemons
      operationId: findAllPokemons
      responses:
        '200':    # status code
          description: sucesso consulta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonDto'

components:
  schemas:
    PokemonDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_default:
          type: boolean
          default: false
        order:
          type: integer
        location_area_encounters:
          type: string
        name:
          type: string
        weight:
          type: integer
        base_experience:
          type: integer
        species:
          $ref: '#/components/schemas/SpeciesDto'
        sprites:
          $ref: '#/components/schemas/SpritesDto'
        abilities:
          type: array
          items:
            $ref: '#/components/schemas/AbilitiesDto'
        forms:
          type: array
          items:
            $ref: '#/components/schemas/FormsDto'
        game_indices:
          type: array
          items:
            $ref: '#/components/schemas/GameIndiceDto'
        held_items:
          type: array
          items:
            $ref: '#/components/schemas/HeldItemDto'
        moves:
          type: array
          items:
            $ref: '#/components/schemas/MoveDto'
        stats:
          type: array
          items:
            $ref: '#/components/schemas/StatsDto'
        types:
          type: array
          items:
            $ref: '#/components/schemas/TypesDto'

    SpeciesDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    SpritesDto:
      type: object
      properties:
        back_default:
          type: string
        front_default:
          type: string

    AbilitiesDto:
      type: object
      properties:
        ability:
          $ref: '#/components/schemas/AbilityDto'
        is_hidden:
          type: boolean
          default: false
        slot:
          type: integer

    AbilityDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    FormsDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    GameIndiceDto:
      type: object
      properties:
        game_index:
          type: integer
        version:
          $ref: '#/components/schemas/VersionDto'

    VersionDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    HeldItemDto:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/ItemDto'
        version_details:
          type: array
          items:
            $ref: '#/components/schemas/VersionDetailsDto'

    VersionDetailsDto:
      type: object
      properties:
        rarity:
          type: integer
        version:
          $ref: '#/components/schemas/VersionDto'

    ItemDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    MoveDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string


    StatsDto:
      type: object
      properties:
        base_stat:
          type: integer
        effort:
          type: integer
        stat:
          $ref: '#/components/schemas/StatDto'

    StatDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string

    TypesDto:
      type: object
      properties:
        slot:
          type: integer
        stat:
          $ref: '#/components/schemas/TypeDto'

    TypeDto:
      type: object
      properties:
        name:
          type: string
        url:
          type: string