# target/generated-sources/src/main/java/com/vegas/swaggercodegen/controller/PokemonApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.30).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.vegas.swaggercodegen.controller;

import com.vegas.swaggercodegen.model.PokemonDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-01-19T12:42:32.059475600-03:00[America/Sao_Paulo]")
@Validated
public interface PokemonApi {

    @Operation(summary = "Obter uma lista de pokemons", description = "Obter uma lista de pokemons", tags={ "Pokemons" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "sucesso consulta", content = @Content(mediaType = "application/json", schema = @Schema(implementation = PokemonDto.class))) })
    @RequestMapping(value = "/pokemon",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<PokemonDto> findAllPokemons();


    @Operation(summary = "Obter um pokemon pelo seu id", description = "Obter um pokemon pelo seu id", tags={ "Pokemons" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "sucesso consulta", content = @Content(mediaType = "application/json", schema = @Schema(implementation = PokemonDto.class))) })
    @RequestMapping(value = "/pokemon/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<PokemonDto> findPokemonById(@Parameter(in = ParameterIn.PATH, description = "ID do pokemon", required=true, schema=@Schema()) @PathVariable("id") Long id);

}